added window and actual export. still buggy
