+++ RailEnginev2/Assets/Editor/ExportWaypoints.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ed295ccc0da415b45993f0f3f3cda80d
-timeCreated: 1443038143
+guid: b0ba3a03e01c1144291020345df40a9b
+timeCreated: 1443038250
 licenseType: Pro
 MonoImporter:
   serializedVersion: 2
