+++ RailEnginev2/Assets/Level.txt
@@ -1,0 +1,13 @@
+enter Level name
+9/24/2015 9:38:49 AM
+enter author name
+M_BEZIER 2 18.65,0,16.61 17,0,0
+M_WAIT 1
+M_BEZIER 10 -1.52,0,39.3 12.1,0,33.4
+M_BEZIER 5 0,0,64.3 -20.3,0,50.4
+M_WAIT 0
+M_MOVE 4 17,0,79
+F_WAIT 5
+F_LOOKAT 2 2 17,0,-25.7 2
+E_WAIT 5
+E_SHAKE 2 0.5
